requirements.txt
Werkzeug==0.16.1
Flask==1.0.2
PyMySQL==0.9.3
Flask-Session==0.3
Jinja2==2.11.2

click==7.1.2
itsdangerous==1.1.0
MarkupSafe==1.1.1
python-dotenv==0.14.0
SQLAlchemy==1.3.19
dominate ==2.5.2 
flask-bootstrap==3.3.7.1
visitor==0.1.3
Flask-SQLAlchemy==2.4.4
flask-marshmallow==0.14.0
flask-RESTful==0.3.8
marshmallow==3.9.0
marshmallow-sqlalchemy==0.24.0
wheel==0.35.1
Flask-Cors==3.0.9